Adding Invoice Endpoints

+++ go-api/models/consumer.go
@@ -1,7 +1,7 @@
 package models
 
 type Consumer struct {
-	ID         int64   `json:"user_id,omitempty"`
+	ID         int     `json:"user_id,omitempty"`
 	FirstName  string  `json:"first_name,omitempty"`
 	LastName   string  `json:"last_name,omitempty"`
 	Email      string  `json:"email,omitempty"`
